+++ test/fixtures/query-auto/nuxt.config.ts
@@ -5,15 +5,7 @@ const resolver = createResolver(import.meta.url)
 const resolve = (...paths: string[]) => resolver.resolve(...paths)
 
 export default defineNuxtConfig({
-  modules: [NuxtEndpoints],
-  endpoints: {
-    client: {
-      query: {
-        setup: 'auto',
-        staleTime: 60_000,
-      },
-    },
-  },
+  modules: ['@pinia/nuxt', '@pinia/colada-nuxt', NuxtEndpoints],
   compatibilityDate: '2024-08-24',
   future: {
     compatibilityVersion: 4,